+++ plugins/wing-map-display/src/map-display/index.js
@@ -7,6 +7,8 @@
 
 import { registerBlockType } from '@wordpress/blocks';
 import Edit from './edit';
+import './style.scss';
+import './editor.scss';
 
 registerBlockType('wing-map-display/wing-map-display', {
 	edit: Edit,